Do not use schema slot grouping; use application-defined custom ordering
